Use strtol (more secure) instead of atoi
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@22151 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/3D/ImageFont.h
@@ -61,7 +61,7 @@ protected:
 
   std::string faceName;
   std::string texture;
-  int	      size;
+  long int    size;
   int	      textureXSize;
   int	      textureYSize;
   int	      textureZStep;
